feat: migrate `MainScroll` and `MainScrollItem` components (#1456)
BREAKING CHANGE: 
`MainScroll` drops the usage of `NoElement` as root element and uses a div instead. This extra div could break the style of an application that relies on `MainScroll` rendering their content directly.
`MainScrollItem` drops the usage of `NoElement` if no `tag` prop is passed and uses a div as fallback.